Deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2
scikit-learn==1.3.2
matplotlib==3.8.2
pandas==2.1.4
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
ultralytics==8.0.206
pyabsa==2.0.4
langchain-groq==0.0.1
langchain-core==0.1.7
langchain==0.1.0
groq==0.3.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1
datasets==2.15.0
tokenizers==0.15.0
scipy==1.11.4
seaborn==0.13.0
tqdm==4.66.1
setuptools==69.0.2
wheel==0.42.0
typing-extensions==4.8.0
httpx==0.25.2
anyio==4.2.0
sniffio==1.3.0
idna==3.6
certifi==2023.11.17
charset-normalizer==3.3.2
urllib3==2.1.0 